requirements.txt
bcrypt
databases[asyncmy]
fastapi
orjson
pydantic
python-dotenv
python-json-logger
pyyaml
redis
uvicorn
